1. If you spend too much time thinking about a thing, you'll never get it done

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -18,7 +18,7 @@
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	size_t a;
+	unsigned int a;
 
 	if (array != NULL || action != NULL)
 	{
